requirements.txt
runpod>=1.7.0
torch==2.1.0
torchvision==0.16.0
opencv-python-headless
numpy<2.0.0
Pillow
modelscope
timm
requests